@@ -27,7 +27,7 @@
     "url": "https://github.com/gribnoysup/react-yandex-maps.git"
   },
   "peerDependencies": {
-    "react": "^0.14.0 || ^15.0.0"
+    "react": "^0.14.9 || ^15.3.0"
   },
   "devDependencies": {
     "babel-cli": "^6.23.0",
@@ -59,10 +59,11 @@
     "lodash": "^4.17.4",
     "prettier": "^0.21.0",
     "pretty-bytes": "^4.0.2",
+    "prop-types": "^15.5.10",
     "raw-loader": "^0.5.1",
-    "react": "^15.4.2",
+    "react": "^15.5.4",
     "react-ace": "^4.1.6",
-    "react-dom": "^15.4.2",
+    "react-dom": "^15.5.4",
     "react-router": "^4.0.0-beta.6",
     "react-router-dom": "^4.0.0-beta.6",
     "source-map-loader": "^0.1.6",
